electron main process refactor
